Move functions to a snake method

+++ snake/make_snake.go
@@ -2,11 +2,11 @@ package snake
 
 import "math/rand"
 
-func MakeSnake(x uint8, y uint8, length uint8) *Snake {
+func MakeSnake(x int32, y int32, length int32) *Snake {
 	snakeBody := make([]SnakePoint, length)
-	snakeBody[0] = SnakePoint{X: uint8(x), Y: uint8(y)}
-	snakeBody[1] = SnakePoint{X: uint8(x), Y: uint8(y - 1)}
-	snakeBody[2] = SnakePoint{X: uint8(x), Y: uint8(y - 2)}
+	snakeBody[0] = SnakePoint{X: x, Y: y}
+	snakeBody[1] = SnakePoint{X: x, Y: int32(y - 1)}
+	snakeBody[2] = SnakePoint{X: x, Y: int32(y - 2)}
 
 	snake := Snake{
 		Body: &snakeBody,
@@ -15,10 +15,10 @@ func MakeSnake(x uint8, y uint8, length uint8) *Snake {
 	return &snake
 }
 
-func MakeSnakeHeadPointUp(x uint8, y uint8, length uint8) *Snake {
+func MakeSnakeHeadPointUp(x int32, y int32, length int32) *Snake {
 	snakeBody := make([]SnakePoint, length)
 	for i := 0; i < int(length); i++ {
-		snakeBody[i] = SnakePoint{X: uint8(x), Y: uint8(y + uint8(i))}
+		snakeBody[i] = SnakePoint{X: x, Y: y + int32(i)}
 	}
 
 	snake := Snake{
@@ -39,11 +39,11 @@ func randSeq(n int) string {
 	return string(b)
 }
 
-func MakeSnakePointingUp(x uint8, y uint8, length uint8) *Snake {
+func MakeSnakePointingUp(x int32, y int32, length int32) *Snake {
 	snakeBody := make([]SnakePoint, length)
 
 	for i := 0; i < int(length); i++ {
-		snakeBody[i] = SnakePoint{X: uint8(x), Y: uint8(y + uint8(i))}
+		snakeBody[i] = SnakePoint{X: int32(x), Y: int32(y + int32(i))}
 	}
 
 	snake := Snake{
@@ -54,11 +54,11 @@ func MakeSnakePointingUp(x uint8, y uint8, length uint8) *Snake {
 	return &snake
 }
 
-func MakeSnakePointingRight(x uint8, y uint8, length uint8) *Snake {
+func MakeSnakePointingRight(x int32, y int32, length int32) *Snake {
 	snakeBody := make([]SnakePoint, length)
 
 	for i := 0; i < int(length); i++ {
-		snakeBody[i] = SnakePoint{X: uint8(x) + uint8(i), Y: uint8(y)}
+		snakeBody[i] = SnakePoint{X: int32(x) + int32(i), Y: int32(y)}
 	}
 
 	snake := Snake{
